Create database. Update coins loading strategy.

+++ app/src/main/java/com/loftschool/loftcoin/domain/ImageUrlFormatter.java
@@ -2,5 +2,5 @@ package com.loftschool.loftcoin.domain;
 
 public interface ImageUrlFormatter {
 
-	String format(int id);
+	String format(long id);
 }
